% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{CxxFlags}
\alias{CxxFlags}
\title{Print RcppBandicoot C++ compiler flags}
\usage{
CxxFlags()
}
\value{
Invisibly returns the flags (after printing them)
}
\description{
Prints the C++ compiler flags required for compiling code that uses RcppBandicoot.
This is a convenience wrapper around \code{\link{RcppBandicootCxxFlags}}.
}
\examples{
CxxFlags()
}
