% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline.R
\name{inlineCxxPlugin}
\alias{inlineCxxPlugin}
\title{Rcpp inline plugin for RcppBandicoot}
\usage{
inlineCxxPlugin()
}
\value{
A list containing the plugin configuration
}
\description{
This function provides the default Rcpp inline plugin for RcppBandicoot.
It uses the configuration determined at package installation time.
}
\examples{
\dontrun{
# Use with Rcpp::sourceCpp()
# File: test.cpp
# // [[Rcpp::depends(RcppBandicoot)]]
#
# #include <RcppBandicoot.h>
#
# // [[Rcpp::export]]
# coot::mat gpu_multiply(const coot::mat& A, const coot::mat& B) {
#   return A * B;
# }
#
# Rcpp::sourceCpp("test.cpp")
}
}
