% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{LdFlags}
\alias{LdFlags}
\title{Print RcppBandicoot linker flags}
\usage{
LdFlags()
}
\value{
Invisibly returns the flags (after printing them)
}
\description{
Prints the linker flags required for linking code that uses RcppBandicoot.
This is a convenience wrapper around \code{\link{RcppBandicootLdFlags}}.
}
\examples{
LdFlags()
}
