% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bandicoot_version}
\alias{bandicoot_version}
\title{Get Bandicoot version}
\usage{
bandicoot_version()
}
\description{
Prints the current version of the Bandicoot library.
}
\examples{
bandicoot_version()
}
