% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_element_square}
\alias{gpu_element_square}
\title{GPU Element-wise Operations}
\usage{
gpu_element_square(A)
}
\arguments{
\item{A}{Input matrix}
}
\value{
Matrix with each element squared, computed on GPU
}
\description{
Apply element-wise square operation on GPU
}
