% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_eye}
\alias{gpu_eye}
\title{Create Identity Matrix on GPU}
\usage{
gpu_eye(n)
}
\arguments{
\item{n}{Size of the identity matrix}
}
\value{
n x n identity matrix on GPU
}
\description{
Create an identity matrix on the GPU
}
\examples{
# See gpu_available()'s example for why this probe is cached and reused
# by every other example in this package.
if (is.null(getOption("rcppbandicoot.ex_info"))) {
  options(rcppbandicoot.ex_info = gpu_device_info())
}
.rcppbandicoot_ex_info <- getOption("rcppbandicoot.ex_info")
if (isTRUE(.rcppbandicoot_ex_info$available)) {
  gpu_eye(3)
}
}
