% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_eye}
\alias{gpu_eye}
\title{Create Identity Matrix on GPU}
\usage{
gpu_eye(n)
}
\arguments{
\item{n}{Size of the identity matrix}
}
\value{
n x n identity matrix on GPU
}
\description{
Create an identity matrix on the GPU
}
