% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_matrix_add}
\alias{gpu_matrix_add}
\title{GPU Matrix Addition}
\usage{
gpu_matrix_add(A, B)
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}
}
\value{
Sum of A and B computed on GPU
}
\description{
Add two matrices on the GPU using Bandicoot
}
