% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_matrix_multiply}
\alias{gpu_matrix_multiply}
\title{GPU Matrix Multiplication}
\usage{
gpu_matrix_multiply(A, B)
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}
}
\value{
Product of A and B computed on GPU
}
\description{
Multiply two matrices on the GPU using Bandicoot
}
