% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_mean}
\alias{gpu_mean}
\title{GPU Mean}
\usage{
gpu_mean(A)
}
\arguments{
\item{A}{Input matrix}
}
\value{
Mean of all elements
}
\description{
Calculate the mean of all elements in a matrix on GPU
}
