% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_randu}
\alias{gpu_randu}
\title{Create Random Matrix on GPU}
\usage{
gpu_randu(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number of rows}

\item{n_cols}{Number of columns}
}
\value{
Random matrix on GPU
}
\description{
Create a matrix with random values on the GPU
}
\examples{
# See gpu_available()'s example for why this probe is cached and reused
# by every other example in this package.
if (is.null(getOption("rcppbandicoot.ex_info"))) {
  options(rcppbandicoot.ex_info = gpu_device_info())
}
.rcppbandicoot_ex_info <- getOption("rcppbandicoot.ex_info")
if (isTRUE(.rcppbandicoot_ex_info$available)) {
  dim(gpu_randu(4, 3))
}
}
