% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_randu}
\alias{gpu_randu}
\title{Create Random Matrix on GPU}
\usage{
gpu_randu(n_rows, n_cols)
}
\arguments{
\item{n_rows}{Number of rows}

\item{n_cols}{Number of columns}
}
\value{
Random matrix on GPU
}
\description{
Create a matrix with random values on the GPU
}
