% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gpu_transpose}
\alias{gpu_transpose}
\title{GPU Matrix Transpose}
\usage{
gpu_transpose(A)
}
\arguments{
\item{A}{Matrix to transpose}
}
\value{
Transposed matrix computed on GPU
}
\description{
Transpose a matrix on the GPU using Bandicoot
}
